forthC

+++ base/views.py
@@ -1,7 +1,12 @@
 from django.shortcuts import render, redirect
-from django.http import HttpResponse
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 # Create your views here.
@@ -13,8 +18,56 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
-    context = {}
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+            print(user)
+        except:
+            messages.error(request, 'User not available!')
+
+        user = authenticate(request, username=username, password=password)
+        # print(user)
+
+        if user is not None:   
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password does not exist!')
+
+    context = {'page':page}
     return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerUser(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration!')
+
+    context = {'form':form}
+    return render(request, 'base/login_register.html', context)
+
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -26,17 +79,33 @@ def home(request):
 
     )
 
+    room_messasges = Message.objects.all()
     room_count = rooms.count()
     topics = Topic.objects.all()
-    context = {'rooms':rooms, 'topics': topics, 'room_count':room_count}
+    context = {'rooms':rooms, 'topics': topics, 'room_count':room_count, 'room_messasges':room_messasges}
     return render(request, 'base/home.html', context)
+
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    rooomMessages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    if request.method == 'POST':
 
+        message  = Message.objects.create(
+            # gets the logged in user
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        # adding the participant to many to many field
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+    context = {'room': room, 'rooomMessages':rooomMessages, 'participants':participants}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -49,9 +118,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
         # print(form)
@@ -62,9 +136,27 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj':message})
